+++ tests/algorithms/imageProcessing/InflectionMapperTwoObjectTest.java
@@ -82,9 +82,9 @@ public class InflectionMapperTwoObjectTest {
 
                 assertNotNull(transformationParams);
 
-                float rotDeg = transformationParams.getRotationInDegrees();
+                double rotDeg = transformationParams.getRotationInDegrees();
 
-                float scale = transformationParams.getScale();
+                double scale = transformationParams.getScale();
                 
                 int nEdges1 = mapper.getEdges1InOriginalReferenceFrame().size();
                 PairIntArray[] edges1 =
